Working implementation of BFS and satcount

+++ include/bdd.h
@@ -60,6 +60,6 @@ bdd_ptr bdd_nor (bdd_ptr a, bdd_ptr b);
 bdd_ptr ithvar(int i);
 
 /* =========== Analysis Functions ========== */
-int countsat(bdd_ptr b);
+double countsat(bdd_ptr b);
 
 #endif /* BDD_H */
